modified test

+++ contracts/util/PropertyOwner.sol
@@ -3,9 +3,9 @@ pragma solidity ^0.4.21;
 import "zeppelin-solidity/contracts/ownership/Ownable.sol";
 
 contract PropertyOwner is Ownable {
-    uint8 deposit = 5;
+    uint deposit = 5;
 
-    function modifyDeposit(uint8 amount) public onlyOwner {
+    function modifyDeposit(uint amount) public onlyOwner {
         deposit = amount;
     }
 }
